+++ spec/content.spec.js
@@ -2,7 +2,7 @@
          prepareEvent, selectElementContents,
          selectElementContentsAndFire,
          placeCursorInsideElement,
-         isFirefox */
+         getEdgeVersion, isFirefox */
 
 describe('Content TestCase', function () {
     'use strict';
@@ -422,6 +422,35 @@ describe('Content TestCase', function () {
             });
             expect(document.execCommand).toHaveBeenCalledWith('unlink', false, null);
         });
+
+        describe('within a blockquote element', function () {
+
+            it('at the end of the blockquote, p tag should be created next, not blockquote', function () {
+                this.el.innerHTML = '<blockquote>lorem ipsum</blockquote>';
+                var editor = this.newMediumEditor('.editor'),
+                    target = editor.elements[0].querySelector('blockquote').firstChild;
+
+                placeCursorInsideElement(target, target.textContent.length);
+                fireEvent(target, 'keydown', {
+                    keyCode: MediumEditor.util.keyCode.ENTER
+                });
+
+                expect(this.el.innerHTML).toBe('<blockquote>lorem ipsum</blockquote><p><br></p>');
+            });
+
+            it('NOT at the start of the blockqoute, no formatting should be changed', function () {
+                this.el.innerHTML = '<blockquote>lorem ipsum</blockquote>';
+                var editor = this.newMediumEditor('.editor'),
+                    target = editor.elements[0].querySelector('blockquote').firstChild;
+
+                placeCursorInsideElement(target, 0);
+                fireEvent(target, 'keydown', {
+                    keyCode: MediumEditor.util.keyCode.ENTER
+                });
+
+                expect(this.el.innerHTML).toBe('<blockquote>lorem ipsum</blockquote>');
+            });
+        });
     });
 
     describe('when the ctrl key and m key is pressed', function () {
@@ -620,7 +649,23 @@ describe('Content TestCase', function () {
                 fireEvent(target, 'keydown', {
                     keyCode: MediumEditor.util.keyCode.BACKSPACE
                 });
-                expect(this.el.innerHTML).toBe('<p>lorem ipsum</p><ul><li></li><li>lorem ipsum</li></ul>');
+                expect(this.el.innerHTML).toMatch(/^<p>lorem ipsum<\/p><ul><li>(<br>)?<\/li><li>lorem ipsum<\/li><\/ul>$/);
+            });
+        });
+
+        describe('within an empty paragraph which is the first element of the editor', function () {
+            it('should delete the paragraph and place the caret to the next paragraph', function () {
+                this.el.innerHTML = '<p class=""><br></p><p>test</p>';
+                var editor = this.newMediumEditor('.editor'),
+                    target = editor.elements[0].querySelector('p'),
+                    range;
+                placeCursorInsideElement(target, 0);
+                fireEvent(target, 'keydown', {
+                    keyCode: MediumEditor.util.keyCode.BACKSPACE
+                });
+                expect(this.el.innerHTML).toBe('<p>test</p>');
+                range = document.getSelection().getRangeAt(0);
+                expect(range.commonAncestorContainer.textContent).toBe('test');
             });
         });
     });
@@ -692,7 +737,7 @@ describe('Content TestCase', function () {
         };
 
         this.el.parentNode.removeChild(this.el);
-        this.el = this.createElement('h1', 'editor', 'M');
+        this.el = this.createElement('h1', 'editor', 'Lorem ipsum dolor sit amet');
 
         var editor = this.newMediumEditor('h1.editor');
         editor.elements[0].focus();
@@ -763,6 +808,77 @@ describe('Content TestCase', function () {
             para = this.el.querySelector('p');
             expect(para.querySelectorAll('br').length).toBe(3, 'Some of the <br> elements have been removed from the <p>');
             expect(para.querySelectorAll('div').length).toBe(0, 'Some <br> elements were replaced with <div> elements within the <p>');
+        });
+    });
+
+    describe('when list element is unlisted', function () {
+        it('should fix markup when one list element is unlisted', function () {
+            this.el.innerHTML = '<ul><li>lorem</li><li>ipsum</li><li>dolor</li></ul>';
+            var editor = this.newMediumEditor('.editor', {
+                    toolbar: {
+                        buttons: ['unorderedlist']
+                    }
+                }),
+                target = editor.elements[0].querySelector('li'),
+                toolbar = editor.getExtensionByName('toolbar');
+
+            selectElementContentsAndFire(target);
+            fireEvent(toolbar.getToolbarElement().querySelector('[data-action="insertunorderedlist"]'), 'click');
+
+            if (getEdgeVersion() > 0) {
+                // Edge wraps elements in div
+                expect(this.el.innerHTML).toBe('<div>lorem</div><ul><li>ipsum</li><li>dolor</li></ul>');
+            } else {
+                // Other browsers should wrap them in p
+                expect(this.el.innerHTML).toBe('<p>lorem</p><ul><li>ipsum</li><li>dolor</li></ul>');
+            }
+        });
+
+        it('should fix markup when miltiple list elements are unlisted', function () {
+            this.el.innerHTML = '<ol><li>lorem</li><li>ipsum</li><li>dolor</li></ol>';
+            var editor = this.newMediumEditor('.editor', {
+                    toolbar: {
+                        buttons: ['orderedlist']
+                    }
+                }),
+                toolbar = editor.getExtensionByName('toolbar'),
+                selection = document.getSelection(),
+                range = document.createRange();
+
+            range.setStart(this.el.querySelectorAll('li')[0].firstChild, 0);
+            range.setEnd(this.el.querySelectorAll('li')[1].firstChild, 5);
+            selection.removeAllRanges();
+            selection.addRange(range);
+
+            fireEvent(toolbar.getToolbarElement().querySelector('[data-action="insertorderedlist"]'), 'click');
+            if (getEdgeVersion() > 0) {
+                // Edge wraps elements in div
+                expect(this.el.innerHTML).toBe('<div>lorem</div><div>ipsum</div><ol><li>dolor</li></ol>');
+            } else {
+                // Other browsers should wrap them in p
+                expect(this.el.innerHTML).toBe('<p>lorem</p><p>ipsum</p><ol><li>dolor</li></ol>');
+            }
+        });
+
+        it('should fix markup when all list elements are unlisted', function () {
+            this.el.innerHTML = '<ul><li>lorem</li><li>ipsum</li><li>dolor</li></ul>';
+            var editor = this.newMediumEditor('.editor', {
+                    toolbar: {
+                        buttons: ['unorderedlist']
+                    }
+                }),
+                target = editor.elements[0].querySelector('ul'),
+                toolbar = editor.getExtensionByName('toolbar');
+
+            selectElementContentsAndFire(target);
+            fireEvent(toolbar.getToolbarElement().querySelector('[data-action="insertunorderedlist"]'), 'click');
+            if (getEdgeVersion() > 0) {
+                // Edge wraps elements in div
+                expect(this.el.innerHTML).toBe('<div>lorem</div><div>ipsum</div><div>dolor</div>');
+            } else {
+                // Other browsers should wrap them in p
+                expect(this.el.innerHTML).toBe('<p>lorem</p><p>ipsum</p><p>dolor</p>');
+            }
         });
     });
 });